requirements.txt
asgiref==3.8.1
diff-match-patch==20230430
Django==5.1
django-import-export==4.1.1
djangorestframework==3.15.2
Faker==27.4.0
psycopg2==2.9.9
python-dateutil==2.9.0.post0
six==1.16.0
sqlparse==0.5.1
tablib==3.5.0
toposort==1.10
tzdata==2024.1
